+++ src/tests/rootsolver.cc
@@ -14,7 +14,7 @@ using namespace smash;
 TEST(root_of_cosine) {
   std::function<double(double)> cosinus = [](double x) { return std::cos(x); };
   auto rootsolver = RootSolver1D(cosinus);
-  double hopefully_pi_half = 0.0;
-  VERIFY(rootsolver.try_find_root(0.0, M_PI, 10000, hopefully_pi_half));
-  COMPARE_ABSOLUTE_ERROR(hopefully_pi_half, M_PI / 2.0, 1e-7);
+  const auto hopefully_pi_half = rootsolver.try_find_root(0.0, M_PI, 10000);
+  VERIFY(hopefully_pi_half.has_value());
+  COMPARE_ABSOLUTE_ERROR((*hopefully_pi_half), M_PI / 2.0, 1e-7);
 }